chore: Update file imports to reflect directory reorganization and update test commands
