+++ package.json
@@ -8,6 +8,8 @@
     "build": "tsc",
     "start": "node dist/index.js",
     "test": "jest --runInBand",
+    "test:sqlite": "DB_TYPE=sqlite jest --runInBand",
+    "test:postgres": "DB_TYPE=postgres jest --runInBand",
     "typeorm": "typeorm"
   },
   "dependencies": {
